♻️ Reorganize API client structure and update import paths
