+++ src/shared/api/index.ts
@@ -1,2 +1,2 @@
-export { apiClient, type ApiClient } from "./api";
-export { queryClient } from "./query-client";
+export { apiClient, type ApiClient } from "./base.api";
+export { queryClient } from "./query-client.api";
